# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Life OS API
  version: 1.0.0
paths:
...
  /v1/tasks/{id}/complete:
    post:
      summary: Mark complete
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Task completed
...
  /v1/focus/sessions/{id}/complete:
    post:
      summary: Complete session
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Focus session completed
...
  /v1/capture:
    post:
      summary: Submit quick capture
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                input_type:
                  type: string
                  enum: [voice, text, photo]
                raw_input:
                  type: string
                audio:
                  type: string
                  format: binary
                photo:
                  type: string
                  format: binary
      responses:
        "200":
          description: Quick capture
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid